perf(signals): 满面板的位置与计数数组从 int64 降到 int32（票据 #78）
## 为什么

全市场选股跑最后一步时偶发整轮失败：

    MemoryError: Unable to allocate 366. MiB for an array with shape (8786, 5453) and data type int64

8786 × 5453 × 8 字节 = 365.5 MiB，也就是一张满面板的形状。面板的行数是「全体交易日的并集」
（本机最早 1990-12-19），不是「这次要用多少根」；而这些数组里存的只是行号与计数，int32
绰绰有余（8,786 行离 2^31 差五个数量级）。

## 改了什么

    signals/filters.py  no_contained_run
      bits（int64，只喂 cumsum）删掉 → cumsum(contained, dtype=int32)
      cumulative / base / positions / latest / left_edge  int64 → int32
    signals/volume.py
      _top_of_advance 的 top、volume_pattern 的 prefix_count  int64 → int32

全是宽度变化，逐格语义不变。

## 实测（同机 A/B，改前用 git stash 取）

    no_contained_run  8786×400   +304 MB → +197 MB   （−35%）
    volume_pattern    8786×400   +707 MB → +654 MB   （−7.5%）
    984 个测试通过、ruff 全绿

端到端（当时手边的 oom_loop 夹具，--cap-gb 12；该夹具随后收进仓库，见 PR #79）：原来死在
过滤器 6/8 的那个分配没了，该过滤器 2.1s 通过；整轮仍然红，但换到了更后面的一个新位置——
volume_pattern 里一次 float64 的转置满面板复制（(5453, 8786)，同样 366 MiB）。那处是下一个
目标。

## 后面那一步（已另开 PR #79）

面板日历截断（把全体交易日的并集切到末端 N 天）——它能把所有面板形状的数组一起降 6.75×，
故上面那个「下一个目标」也一并消失。它动的是 ADR-0014 记下的决策，按流程先 grill 过，故
单独提一个 PR。

+++ src/mbt/signals/volume.py
@@ -463,7 +463,8 @@ def _top_of_advance(high: np.ndarray, trough_pos: np.ndarray) -> np.ndarray:
     推进，故重扫的总代价是 O(行数·列数)，不是 O(段长·行数)。实测全市场约 4 秒。
     """
     rows, columns = high.shape
-    top = np.full((rows, columns), -1, dtype=np.int64)
+    # 存的是**行号**（-1 表示段无效），不是价格：int32 足够（见 issue #78 那条注记）。
+    top = np.full((rows, columns), -1, dtype=np.int32)
     for column in range(columns):
         current = -1
         best = -1.0
@@ -673,8 +674,9 @@ def volume_pattern(
     finite = np.isfinite(vol)
     prefix_total = np.zeros((rows + 1, columns), dtype=float)
     prefix_total[1:] = np.where(finite, vol, 0.0).cumsum(axis=0)
-    prefix_count = np.zeros((rows + 1, columns), dtype=np.int64)
-    prefix_count[1:] = finite.cumsum(axis=0)
+    # 计数是**窗口内有限值的个数**，最大就是行数，int32 足够（见 issue #78 那条注记）。
+    prefix_count = np.zeros((rows + 1, columns), dtype=np.int32)
+    prefix_count[1:] = finite.cumsum(axis=0, dtype=np.int32)
 
     for column in range(columns):
         first, last = int(bounds[column]), int(bounds[column + 1])
